refactor: moved access checkers to separate top-level package
