+++ src/main/kotlin/se/uulm/snowballr/backend/access/AccessChecker.kt
@@ -1,4 +1,4 @@
-package se.uulm.snowballr.backend.service.accessrules
+package se.uulm.snowballr.backend.access
 
 /**
  * The [IAccessChecker] interface defines the contract for access checking in the SnowballR backend.
