+++ src/main/kotlin/se/uulm/snowballr/backend/access/rules/AccessRule.kt
@@ -1,4 +1,4 @@
-package se.uulm.snowballr.backend.service.accessrules
+package se.uulm.snowballr.backend.access.rules
 
 import se.uulm.snowballr.backend.model.dto.User
 import se.uulm.snowballr.backend.model.exception.SnowballRException
