+++ src/main/kotlin/se/uulm/snowballr/backend/service/CriterionService.kt
@@ -1,5 +1,7 @@
 package se.uulm.snowballr.backend.service
 
+import se.uulm.snowballr.backend.access.ICriterionAccessChecker
+import se.uulm.snowballr.backend.access.IProjectAccessChecker
 import se.uulm.snowballr.backend.grpc.SnowballRServer.SnowballRService
 import se.uulm.snowballr.backend.model.EntityType
 import se.uulm.snowballr.backend.model.dto.toGrpcCriteria
@@ -7,8 +9,6 @@ import se.uulm.snowballr.backend.model.dto.toGrpcCriterion
 import se.uulm.snowballr.backend.model.parseUUID
 import se.uulm.snowballr.backend.repository.ICriterionTableRepo
 import se.uulm.snowballr.backend.repository.IUserTableRepo
-import se.uulm.snowballr.backend.service.accessrules.ICriterionAccessChecker
-import se.uulm.snowballr.backend.service.accessrules.IProjectAccessChecker
 import java.util.UUID
 import snowballr.CriterionOuterClass.Criterion as GrpcCriterion
